+++ src/FatturaElettronica/FatturaElettronicaBody/DatiBeniServizi/Linea.php
@@ -11,13 +11,14 @@
 
 namespace Deved\FatturaElettronica\FatturaElettronica\FatturaElettronicaBody\DatiBeniServizi;
 
+use Deved\FatturaElettronica\Enum\NaturaIvaType;
 use Deved\FatturaElettronica\Traits\MagicFieldsTrait;
 use Deved\FatturaElettronica\XmlSerializableInterface;
 
 class Linea implements XmlSerializableInterface
 {
     use MagicFieldsTrait;
-    /** @var integer */
+    /** @var int */
     protected $numeroLinea;
     /** @var string */
     protected $codiceArticolo;
@@ -31,6 +32,16 @@ class Linea implements XmlSerializableInterface
     protected $prezzoUnitario;
     /** @var float */
     protected $aliquotaIva;
+    /** @var string */
+    protected $codiceTipo;
+    /** @var ScontoMaggiorazione[]|null */
+    protected $scontoMaggiorazione = [];
+    /** @var int */
+    protected $decimaliLinea;
+    /** @var string */
+    protected $tipoCessionePrestazione;
+    /** @var NaturaIvaType|null */
+    protected $naturaIva;
 
 
     /**
@@ -41,6 +52,8 @@ class Linea implements XmlSerializableInterface
      * @param float $quantita
      * @param string $unitaMisura
      * @param float $aliquotaIva
+     * @param string $codiceTipo
+     * @param int $decimaliLinea
      */
     public function __construct(
         $descrizione,
@@ -48,7 +61,11 @@ class Linea implements XmlSerializableInterface
         $codiceArticolo = null,
         $quantita = null,
         $unitaMisura = 'pz',
-        $aliquotaIva = 22.00
+        $aliquotaIva = 22.00,
+        $codiceTipo = 'FORN',
+        $decimaliLinea = 2,
+        $tipoCessionePrestazione = null,
+        $naturaIva = null 
     ) {
         $this->codiceArticolo = $codiceArticolo;
         $this->descrizione = $descrizione;
@@ -56,6 +73,10 @@ class Linea implements XmlSerializableInterface
         $this->quantita = $quantita;
         $this->unitaMisura = $unitaMisura;
         $this->aliquotaIva = $aliquotaIva;
+        $this->codiceTipo = $codiceTipo;
+        $this->decimaliLinea = $decimaliLinea;
+        $this->tipoCessionePrestazione = $tipoCessionePrestazione;
+        $this->naturaIva = $naturaIva;
     }
 
 
@@ -67,23 +88,31 @@ class Linea implements XmlSerializableInterface
     {
         $writer->startElement('DettaglioLinee');
         $writer->writeElement('NumeroLinea', $this->numeroLinea);
+        if ($this->tipoCessionePrestazione) {
+            $writer->writeElement('TipoCessionePrestazione', $this->tipoCessionePrestazione);
+        }
         if ($this->codiceArticolo) {
             $writer->startElement('CodiceArticolo');
-                $writer->writeElement('CodiceTipo', 'FORN');
-                //todo: implementare altri tipi di codice
-                $writer->writeElement('CodiceValore', $this->codiceArticolo);
+            $writer->writeElement('CodiceTipo', $this->codiceTipo);
+            $writer->writeElement('CodiceValore', $this->codiceArticolo);
             $writer->endElement();
         }
         $writer->writeElement('Descrizione', $this->descrizione);
         if ($this->quantita) {
-            $writer->writeElement('Quantita', fe_number_format($this->quantita, 2));
+            $writer->writeElement('Quantita', fe_number_format($this->quantita, $this->decimaliQuantita()));
             $writer->writeElement('UnitaMisura', $this->unitaMisura);
         }
         $this->writeXmlField('DataInizioPeriodo', $writer);
         $this->writeXmlField('DataFinePeriodo', $writer);
-        $writer->writeElement('PrezzoUnitario', fe_number_format($this->prezzoUnitario, 2));
+        $writer->writeElement('PrezzoUnitario', fe_number_format($this->prezzoUnitario, $this->decimaliLinea));
+        foreach ($this->scontoMaggiorazione as $item) {
+            $item->toXmlBlock($writer);
+        }
         $writer->writeElement('PrezzoTotale', $this->prezzoTotale());
         $writer->writeElement('AliquotaIVA', fe_number_format($this->aliquotaIva, 2));
+        if($this->naturaIva !== null) {
+            $writer->writeElement('Natura', $this->naturaIva);
+        }    
         $this->writeXmlFields($writer);
         $writer->endElement();
         return $writer;
@@ -97,11 +126,25 @@ class Linea implements XmlSerializableInterface
      */
     public function prezzoTotale($format = true)
     {
-        $quantita = $this->quantita ? $this->quantita : 1;
+        $quantita = $this->quantita ?: 1;
+        $totale = $this->prezzoUnitario * $quantita;
         if ($format) {
-            return fe_number_format($this->prezzoUnitario * $quantita, 2);
+            $totale = fe_number_format($totale, $this->decimaliLinea);
         }
-        return $this->prezzoUnitario * $quantita;
+        foreach ($this->scontoMaggiorazione as $item) {
+            $totale = $item->applicaScontoMaggiorazione($totale, $quantita, $format ? $this->decimaliLinea : null);
+        }
+        return fe_number_format($totale, $this->decimaliLinea);
+    }
+
+    /**
+     * Restituisce il numero di decimali della quantita
+     *
+     * @return int
+     */
+    public function decimaliQuantita()
+    {
+        return max(min(strlen(substr(strrchr($this->quantita, "."), 1)), 8), 2);
     }
 
     /**
@@ -122,5 +165,10 @@ class Linea implements XmlSerializableInterface
     public function getAliquotaIva()
     {
         return $this->aliquotaIva;
+    }
+
+    public function setScontoMaggiorazione(ScontoMaggiorazione $scontoMaggiorazione)
+    {
+        $this->scontoMaggiorazione[] = $scontoMaggiorazione;
     }
 }